requirements.txt
matplotlib==3.1.2
pandas==0.25.2
Pillow==7.0.0
requests==2.22.0